fix(security): prevent XSS in portfolio generation by sanitizing URLs
Refactored `generatePortfolioHtml` into a dedicated library `server/lib/portfolio-generator.ts`.
Added `sanitizeUrl` to block dangerous protocols (`javascript:`, `data:`, `vbscript:`) in repository URLs.
Added `tests/portfolio-security.test.ts` to verify the security fix.
Removed local implementation of portfolio generation from `server/routes/deploy.ts`.
